+++ src/lib/components/WhiteTextBox.svelte
@@ -2,7 +2,7 @@
 </script>
 
 <div
-	class="dark:bg-primary-900 mt-12 flex max-w-[90%] flex-col rounded-md bg-white p-4 text-black dark:text-white md:mt-4 md:max-w-[50%] md:p-12">
+	class="dark:bg-random-970 mt-12 flex max-w-[90%] flex-col rounded-md bg-white p-4 text-black dark:text-white md:mt-4 md:max-w-[50%] md:p-12">
 	<slot />
 </div>