+++ src/routes/+layout.ts
@@ -1,0 +1,8 @@
+export const prerender = true;
+export const ssr = true;
+
+/** @type {import('@sveltejs/kit').Load} */
+export const load = async () => {
+
+	return { };
+};