+++ src/routes/projects/+page.js
@@ -1,6 +1,0 @@
-/** @type {import('./$types').PageLoad} */
-import json from "/src/lib/projectList.json";
-
-export async function load({ fetch }) {
-	return { projects: json };
-}