+++ svelte.config.js
@@ -1,17 +1,14 @@
 // use adapter-auto for zero config deployment to Vercel, adapter-static for Github Pages
 // import adapter from '@sveltejs/adapter-auto';
 // import preprocess from "svelte-preprocess";
-// import { vitePreprocess } from "@sveltejs/kit/vite";
 import adapter from "@sveltejs/adapter-static";
-import { vitePreprocess } from "@sveltejs/kit/vite";
-import { preprocessMeltUI } from "@melt-ui/pp";
-import sequence from "svelte-sequential-preprocessor";
+import preprocess from "svelte-preprocess";
 
 const dev = process.argv.includes("dev");
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-	preprocess: sequence([vitePreprocess(), preprocessMeltUI()]),
+	preprocess: preprocess(),
 	kit: {
 		adapter: adapter(),
 	},